BUG: need to provide int64_t support for platforms with 4 byte longs

+++ AIOUSB/samples/TestLib/aiocommon.c
@@ -204,7 +204,7 @@ void process_aio_cmd_line( struct opts *options, int argc, char *argv [] )
             break;
         case 'N':
         case 'b':
-            options->num_scans = atol(optarg);
+            options->num_scans = (int64_t)atoll(optarg);
             if( options->num_scans <= 0 ) {
                 fprintf(stderr,"Warning: Buffer Size outside acceptable range (1,1e8), setting to 10000\n");
                 options->num_scans = 10000;
